+++ examples/python/ghz.py
@@ -12,39 +12,27 @@ Note: if you have only cloned the Qiskit repository but not
 used `pip install`, the examples only work from the root directory.
 """
 
-from qiskit import QuantumProgram
+from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
+from qiskit import register, execute
 import Qconfig
 
 ###############################################################
 # Set the backend name and coupling map.
 ###############################################################
-backend = "ibmqx2"
-coupling_map = {0: [1, 2],
-                1: [2],
-                2: [],
-                3: [2, 4],
-                4: [2]}
+backend = "ibmq_5_tenerife"
+coupling_map = [[0, 1],
+                [0, 2],
+                [1, 2],
+                [3, 2],
+                [3, 4],
+                [4, 2]]
 
 ###############################################################
-# Make a quantum program for the GHZ state.
+# Make a quantum circuit for the GHZ state.
 ###############################################################
-QPS_SPECS = {
-    "circuits": [{
-        "name": "ghz",
-        "quantum_registers": [{
-            "name": "q",
-            "size": 5
-        }],
-        "classical_registers": [
-            {"name": "c",
-             "size": 5}
-        ]}]
-}
-
-qp = QuantumProgram(specs=QPS_SPECS)
-qc = qp.get_circuit("ghz")
-q = qp.get_quantum_register("q")
-c = qp.get_classical_register("c")
+q = QuantumRegister(5, "q")
+c = ClassicalRegister(5, "c")
+qc = QuantumCircuit(q, c, name='ghz')
 
 # Create a GHZ state
 qc.h(q[0])
@@ -59,32 +47,32 @@ for i in range(5):
 ###############################################################
 # Set up the API and execute the program.
 ###############################################################
-qp.set_api(Qconfig.APItoken, Qconfig.config["url"])
+register(Qconfig.APItoken, Qconfig.config["url"])
 
 # First version: no mapping
-print("no mapping, simulator")
-result = qp.execute(["ghz"], backend='ibmq_qasm_simulator',
-                    coupling_map=None, shots=1024)
+print("no mapping, execute on simulator")
+job = execute(qc, backend='ibmq_qasm_simulator', coupling_map=None, shots=1024)
+result = job.result()
 print(result)
 print(result.get_counts("ghz"))
 
-# Second version: map to qx2 coupling graph and simulate
-print("map to %s, simulator" % backend)
-result = qp.execute(["ghz"], backend='ibmq_qasm_simulator',
-                    coupling_map=coupling_map, shots=1024)
+# Second version: map to ibmq_5_tenerife coupling graph and simulate online
+print("map to %s, execute on online simulator" % backend)
+job = execute(qc, backend='ibmq_qasm_simulator', coupling_map=coupling_map, shots=1024)
+result = job.result()
 print(result)
 print(result.get_counts("ghz"))
 
-# Third version: map to qx2 coupling graph and simulate locally
-print("map to %s, local qasm simulator" % backend)
-result = qp.execute(["ghz"], backend='local_qasm_simulator',
-                    coupling_map=coupling_map, shots=1024)
+# Third version: map to ibmq_5_tenerife coupling graph and simulate locally
+print("map to %s, execute on local simulator" % backend)
+job = execute(qc, backend='local_qasm_simulator', coupling_map=coupling_map, shots=1024)
+result = job.result()
 print(result)
 print(result.get_counts("ghz"))
 
-# Fourth version: map to qx2 coupling graph and run on qx2
-print("map to %s, backend" % backend)
-result = qp.execute(["ghz"], backend=backend,
-                    coupling_map=coupling_map, shots=1024, timeout=120)
+# Fourth version: map to ibmq_5_tenerife coupling graph and run on ibmq_5_tenerife
+print("map to %s, execute on device" % backend)
+job = execute(qc, backend=backend, shots=1024)
+result = job.result()
 print(result)
 print(result.get_counts("ghz"))